Implement basic dead cell check
Add functionality to check neighbors of a given cell. Use this
functionality to check for each generation if a dead cell should come to
life.

Functionality for modifying dead cells to come to life is incomplete.
Cells need to be modified simultaneously so as not to interfere with the
neighbor-checking logic of subsequenty checked cells.